Updated Zoom Recording type, splitting it into two types based on originator (user/meeting)
