+++ lib/types/ZoomMeetingOrWebinarType.d.ts
@@ -1,5 +1,8 @@
 /**
- * Types of Zoom meetings
+ * Types of Zoom meetings. This is a combination of multiple types in the Zoom
+ *   API and can be found by combining fields from parts of the Zoom Meeting
+ *   settings object {@link https://marketplace.zoom.us/docs/api-reference/zoom-api/methods/#operation/meetings}
+ *   and the Zoom Webinar settings object {@link https://marketplace.zoom.us/docs/api-reference/zoom-api/methods/#operation/webinars}
  * @author Gabe Abrams
  */
 declare enum ZoomMeetingOrWebinarType {
