+++ lib/types/ZoomMeetingOrWebinarType.js
@@ -1,7 +1,10 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
 /**
- * Types of Zoom meetings
+ * Types of Zoom meetings. This is a combination of multiple types in the Zoom
+ *   API and can be found by combining fields from parts of the Zoom Meeting
+ *   settings object {@link https://marketplace.zoom.us/docs/api-reference/zoom-api/methods/#operation/meetings}
+ *   and the Zoom Webinar settings object {@link https://marketplace.zoom.us/docs/api-reference/zoom-api/methods/#operation/webinars}
  * @author Gabe Abrams
  */
 var ZoomMeetingOrWebinarType;
