+++ lib/types/ZoomUserRecordings.d.ts
@@ -1,0 +1,15 @@
+import ZoomMeetingRecordings from './ZoomMeetingRecordings';
+/**
+ * List of recordings for a user
+ * @author Gabe Abrams
+ */
+declare type ZoomUserRecordings = {
+    from: string;
+    to: string;
+    next_page_token: string;
+    page_count: number;
+    page_size: number;
+    total_records: number;
+    meetings: ZoomMeetingRecordings[];
+};
+export default ZoomUserRecordings;
