--- src/types/ZoomCloudRecording.ts
@@ -1,9 +1,0 @@
-/**
- * Zoom cloud recording object {@link https://marketplace.zoom.us/docs/api-reference/zoom-api/methods/#operation/recordingGet}
- * @author Gabe Abrams
- */
-type ZoomCloudRecording = {
-  
-};
-
-export default ZoomCloudRecording;
